Added startup sequence checking. Fixed PtP route not being included in the EOT and UPDATE packets.

+++ eigrpd/eigrp_structs.h
@@ -242,7 +242,7 @@ struct eigrp_neighbor {
 	struct eigrp_interface *ei;
 
 	/* EIGRP neighbor Information */
-	uint8_t state; /* neigbor status. */
+	uint32_t state; /* neigbor status. */
 
 	uint32_t recv_sequence_number; /* Last received sequence Number. */
 	uint32_t init_sequence_number;
